LPS-56863 Move classes only
