+++ modules/apps/portal-instances/portal-instances-web/src/com/liferay/portal/instances/web/portlet/action/ViewMVCRenderCommand.java
@@ -12,16 +12,12 @@
  * details.
  */
 
-package com.liferay.portlet.portalinstances.action;
+package com.liferay.portal.instances.web.portlet.action;
 
-import com.liferay.portal.NoSuchCompanyException;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
-import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
-import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.util.PortletKeys;
 
-import javax.portlet.PortletException;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
@@ -31,33 +27,16 @@ import javax.portlet.RenderResponse;
 @OSGiBeanProperties(
 	property = {
 		"javax.portlet.name=" + PortletKeys.PORTAL_INSTANCES,
-		"mvc.command.name=/portal_instances/edit_instance"
+		"mvc.command.name=/", "mvc.command.name=/portal_instances/view"
 	}
 )
-public class EditInstanceMVCRenderCommand implements MVCRenderCommand {
+public class ViewMVCRenderCommand implements MVCRenderCommand {
 
 	@Override
 	public String render(
-			RenderRequest renderRequest, RenderResponse renderResponse)
-		throws PortletException {
+		RenderRequest renderRequest, RenderResponse renderResponse) {
 
-		try {
-			ActionUtil.getInstance(renderRequest);
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchCompanyException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(renderRequest, e.getClass());
-
-				return "/html/portlet/portal_instances/error.jsp";
-			}
-			else {
-				throw new PortletException(e);
-			}
-		}
-
-		return "/html/portlet/portal_instances/edit_instance.jsp";
+		return "/html/portlet/portal_instances/view.jsp";
 	}
 
 }
